+++ wdl-lint/src/rules/unknown_rule.rs
@@ -42,8 +42,9 @@ impl Rule for UnknownRule {
     }
 
     fn explanation(&self) -> &'static str {
-        "When writing WDL, lint directives are used to suppress certain rules. If a rule is unknown, \
-         nothing will be suppressed. This rule flags unknown rules as they are often mistakes."
+        "When writing WDL, lint directives are used to suppress certain rules. If a rule is \
+         unknown, nothing will be suppressed. This rule flags unknown rules as they are often \
+         mistakes."
     }
 
     fn tags(&self) -> TagSet {